Worked on cargo support on the ship and prepared work for the route and islands
Added an enum for stats as it allows adding a None variable easier than using strings and give more structure. Provided the basic support to add, see, and remove items from the ship changing the stats as it does so. One of the next steps is to add in error checking so no value can be less than 0 and the capacity can't go under what it is carrying.